Easy fix for #4483
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/trunk/kdchart@10463 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ src/KDChartCartesianAxis.cpp
@@ -221,7 +221,7 @@ void CartesianAxis::Private::drawSubUnitRulers( QPainter* painter, CartesianCoor
     int nextMayBeTick = 0;
     int mayBeTick = 0;
     int logSubstep = 0;
-    float f = dim.start;
+    qreal f = dim.start;
     qreal fLogSubstep = f;
     const bool isAbscissa = axis()->isAbscissa();
     const bool isLogarithmic = (dim.calcMode == AbstractCoordinatePlane::Logarithmic );
